Store passwords as String
Reason to do so is that the tmp-file, while editing, will display
passwords in plaintext, so they are easy to edit.

The reasons for not doing so, AFAIK, is that since String is immutable,
a memory dump would expose the password. However, I can't see a big risk
for a memory dump being taken. A bigger concern would be the saved
temporary file, since it is written to disk and will exist there until
overwritten.

An attacker able to get hold of the temporary file would probably also
have the competency to recognize the stored password as character bytes
and convert them, so I don't consider it "added security" either.